Move linkpreview to messenger
